Organize the tests into modules loosely bundled as features
