+++ test/utils/cleandb.sh
